CLEANUP of old bash scripts, imprroved deploy

+++ scripts/remote/phpmyadmin/install.js
@@ -21,14 +21,27 @@ const phpmyadminInstall = async() => {
   const tmp = `${SERVER_PATH_CODEBASE}/.tmp/phpmyadmin`;
   const dest = `${SERVER_PATH_WEBROOT}/phpmyadmin`;
 
+  if(await ssh.directoryExists(dest)) {
+    logger.warning(`Path ${dest} already exists.`);
+    logger.warning('Aborting.');
+    return;
+  }
+
   await ssh.exec(`mkdir -p ${tmp}`);
   await ssh.exec(`curl -s -L ${sourceUrl} --output ${tmp}/phpmyadmin.tar.gz`);
   await ssh.exec(`tar -zxf ${tmp}/phpmyadmin.tar.gz --directory ${tmp}`);
-  await ssh.exec(`mv -f ${tmp}/${dist} ${dest}`);
+  await ssh.exec(`mv -f ${tmp}/phpMyAdmin-${dist} ${dest}`);
   await ssh.exec(`chown -R ${WEB_USER_NAME}:${WEB_USER_GROUP} ${dest}`);
   await ssh.exec(`rm -rf ${tmp}`);
 
-  logger.info(`phpMyAdmin ${dist} is installed!`);
+  logger.success(`phpMyAdmin ${dist} is installed!`);
 };
 
 module.exports = phpmyadminInstall;
+
+phpmyadminInstall()
+  .then(process.exit)
+  .catch(err => {
+    console.error(err);
+    process.exit(1);
+  });
